+++ apps/core/models/task.py
@@ -42,7 +42,7 @@ class Task(BaseModel):
         return reverse("task-detail", kwargs={"task_id": self.pk})
 
     def __str__(self):
-        return f"{self.assigment.name} - {self.user.username} ({self.id})"
+        return f"{self.assigment.name} - ({self.id})"
 
 
 __all__ = ["Task"]