+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.11.0 : 2024-02-08
 
 - **Added**: [Save and prefill image on new test #5](https://github.com/FIIT-Databases/tester/issues/5)
+- **Added**: Cool FIIT favicon (super important)
 - **Changed**: Django 5.0 upgrade
 - **Changed**: Use Debian-based containers