+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **Added**: [Save and prefill image on new test #5](https://github.com/FIIT-Databases/tester/issues/5)
 - **Added**: Cool FIIT favicon (super important)
+- **Added**: Advanced admin filters
 - **Changed**: Django 5.0 upgrade
 - **Changed**: Use Debian-based containers